Conflicts:
	tajo-catalog/tajo-catalog-common/src/main/java/org/apache/tajo/catalog/TableDesc.java
	tajo-catalog/tajo-catalog-server/src/main/java/org/apache/tajo/catalog/store/MemStore.java
	tajo-storage/tajo-storage-hdfs/src/main/java/org/apache/tajo/storage/FileAppender.java

+++ tajo-plan/src/main/java/org/apache/tajo/plan/expr/WindowFunctionEval.java
@@ -65,6 +65,9 @@ public class WindowFunctionEval extends AggregationFunctionCallEval implements C
 
   @Override
   public Datum terminate(FunctionContext context) {
+    if (!isBinded) {
+      throw new IllegalStateException("bind() must be called before terminate()");
+    }
     return instance.terminate(context);
   }
